+++ tag/writer.py
@@ -15,8 +15,9 @@ except ImportError:  # pragma: no cover
     from io import StringIO
 import sys
 import tag
-from tag.feature import Feature
-from tag.reader import GFF3Reader
+from tag import Feature
+from tag import Sequence
+from tag import GFF3Reader
 
 
 class GFF3Writer():
@@ -45,12 +46,13 @@ class GFF3Writer():
             self.outfile = tag.open(outfile, 'w')
         self.retainids = False
         self.feature_counts = defaultdict(int)
+        self._seq_written = False
 
     def __del__(self):
         if self.outfilename != '-' and not isinstance(self.outfile, StringIO):
             self.outfile.close()
 
-    def write(self, relax=False):
+    def write(self):
         """Pull features from the instream and write them to the output."""
         for entry in self._instream:
             if isinstance(entry, Feature):
@@ -62,4 +64,7 @@ class GFF3Writer():
                         feature.add_attribute('ID', fid)
                     else:
                         feature.drop_attribute('ID')
+            if isinstance(entry, Sequence) and not self._seq_written:
+                print('##FASTA', file=self.outfile)
+                self._seq_written = True
             print(repr(entry), file=self.outfile)