+++ tests/test_select.py
@@ -9,28 +9,26 @@
 
 import pytest
 import tag
+from tag import GFF3Reader
 
 
 def test_features():
-    ifn = 'tests/testdata/pbar-withseq.gff3'
-
-    reader = tag.reader.GFF3Reader(infilename=ifn)
+    reader = GFF3Reader(infilename='tests/testdata/pbar-withseq.gff3')
     features = [f for f in tag.select.features(reader)]
     assert len(features) == 3
 
-    reader = tag.reader.GFF3Reader(infilename=ifn)
+    reader = GFF3Reader(infilename='tests/testdata/pbar-withseq.gff3')
     features = [f for f in tag.select.features(reader, type='pseudogene')]
     assert len(features) == 1
 
-    reader = tag.reader.GFF3Reader(infilename=ifn)
+    reader = GFF3Reader(infilename='tests/testdata/pbar-withseq.gff3')
     with pytest.raises(ValueError) as ve:
         features = [f for f in tag.select.features(reader, traverse=True)]
     assert 'cannot traverse without a specific feature type' in str(ve)
 
 
 def test_window_contain():
-    ifn = 'tests/testdata/pdom-withseq.gff3'
-    reader = tag.reader.GFF3Reader(infilename=ifn)
+    reader = GFF3Reader(infilename='tests/testdata/pdom-withseq.gff3')
     features = tag.select.features(reader)
     window = tag.select.window(features, 'PdomSCFr1.2-0483', 3000, 7000)
     selected = [f for f in window]
@@ -39,8 +37,7 @@ def test_window_contain():
 
 
 def test_window_overlap():
-    ifn = 'tests/testdata/osat-twoscaf.gff3.gz'
-    reader = tag.reader.GFF3Reader(infilename=ifn)
+    reader = GFF3Reader(infilename='tests/testdata/osat-twoscaf.gff3.gz')
     features = tag.select.features(reader)
     window = tag.select.window(features, 'NW_015379189.1', 15000, 20000,
                                strict=False)
@@ -50,8 +47,7 @@ def test_window_overlap():
 
 
 def test_window_seqid():
-    ifn = 'tests/testdata/osat-twoscaf.gff3.gz'
-    reader = tag.reader.GFF3Reader(infilename=ifn)
+    reader = GFF3Reader(infilename='tests/testdata/osat-twoscaf.gff3.gz')
     features = tag.select.features(reader)
     window = tag.select.window(features, 'NW_015379208.1')
     selected = [f for f in window]
@@ -60,22 +56,18 @@ def test_window_seqid():
 
 
 def test_directives():
-    ifn = 'tests/testdata/pbar-withseq.gff3'
-
-    reader = tag.reader.GFF3Reader(infilename=ifn)
+    reader = GFF3Reader(infilename='tests/testdata/pbar-withseq.gff3')
     directives = [d for d in tag.select.directives(reader)]
     assert len(directives) == 3
 
-    reader = tag.reader.GFF3Reader(infilename=ifn)
+    reader = GFF3Reader(infilename='tests/testdata/pbar-withseq.gff3')
     dirs = [d for d in tag.select.directives(reader, type='sequence-region')]
     assert len(dirs) == 2
     assert [d.seqid for d in dirs] == ['NW_011929623.1', 'NW_011929624.1']
 
 
 def test_sequences():
-    ifn = 'tests/testdata/pbar-withseq.gff3'
-
-    reader = tag.reader.GFF3Reader(infilename=ifn)
+    reader = GFF3Reader(infilename='tests/testdata/pbar-withseq.gff3')
     seqs = [s for s in tag.select.sequences(reader)]
     assert len(seqs) == 2
     assert sorted([s.seq[0:5] for s in seqs]) == ['GCTAA', 'TAGAC']