+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- An index class for efficient in-memory access of sequence features.
+
 ## [0.1.1] - 2016-12-19
 ### Changed
 - CLI implemented using `entry_points` instead of a dedicated script.