+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- Module for mRNA handling, with a function for selecting the primary mRNA from
+  a gene or other feature
+
 ## [0.1.1] - 2016-12-19
 ### Changed
 - CLI implemented using `entry_points` instead of a dedicated script.