+++ CHANGELOG.md
@@ -3,7 +3,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-## Changed
+### Added
+- Module for mRNA handling, with a function for selecting the primary mRNA from
+  a gene or other feature.
+- New CLI command `tag pmrna`.
+
+### Changed
 - Modules focused on classes / data structure now support more concise imports
   (for example, `from tag import Feature` and `tag.Feature` now supported and
   preferred over `from tag.feature import feature` and `tag.feature.Feature`).